Actor spawner rules; various build fixes; slimes tests; prototype
