+++ Assets/Scripts/Development/Game/Level/ALevelParams.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5c14ab7ccb48c4144a5797a7a04be483
+timeCreated: 1475183031
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
