+++ Assets/Scripts/Development/Game/Level/Tiled/Dungeon/Map/ActorSpawnerData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3fc308fd82ef7774ca77a947d4836c8f
+timeCreated: 1475184982
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
