+++ Assets/Scripts/Development/Game/Level/Tiled/_Map/MapCollision.cs
@@ -36,7 +36,7 @@ namespace Game.Level.Tiled
 		{
 			BuildColliders();
 
-			Built();
+			Built(GetType());
 		}
 
 		public void DestroyColliders()
