--- Assets/Scripts/Prototype.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: bd690ef8c83226a469935d7a3abc2d0e
-folderAsset: yes
-timeCreated: 1474470595
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
